SslTest: split `verify_reference_name()` specific tests to dedicated test class
