+++ tests/Ssl/SslTest.php
@@ -404,100 +404,6 @@ final class SslTest extends TestCase {
 	}
 
 	/**
-	 * Test correctly identifying whether a reference name is valid.
-	 *
-	 * @covers ::verify_reference_name
-	 *
-	 * @dataProvider dataVerifyReferenceName
-	 *
-	 * @param string $reference Reference name to test.
-	 * @param bool   $expected  Expected function outcome.
-	 *
-	 * @return void
-	 */
-	public function testVerifyReferenceName($reference, $expected) {
-		$this->assertSame($expected, Ssl::verify_reference_name($reference));
-	}
-
-	/**
-	 * Data provider.
-	 *
-	 * @return array
-	 */
-	public function dataVerifyReferenceName() {
-		return [
-			'empty string' => [
-				'reference' => '',
-				'expected'  => false,
-			],
-			'one part, no dot' => [
-				'reference' => 'example',
-				'expected'  => true,
-			],
-			'one part, only wildcard' => [
-				'reference' => '*',
-				'expected'  => false,
-			],
-			'two parts, no wildcard' => [
-				'reference' => 'example.com',
-				'expected'  => true,
-			],
-			'two parts, wildcard in first' => [
-				'reference' => '*.com',
-				'expected'  => false,
-			],
-			'two parts, wildcard in last' => [
-				'reference' => 'example.*',
-				'expected'  => false,
-			],
-			'three parts, only dots' => [
-				'reference' => '..',
-				'expected'  => false,
-			],
-			'three parts, no wildcard' => [
-				'reference' => new StringableObject('www.example.com'),
-				'expected'  => true,
-			],
-			'three parts, no wildcard, has spaces' => [
-				'reference' => 'my dog . and . my cat',
-				'expected'  => false,
-			],
-			'three parts, wildcard in first' => [
-				'reference' => '*.example.com',
-				'expected'  => true,
-			],
-			'three parts, wildcard in second' => [
-				'reference' => 'www.*.com',
-				'expected'  => false,
-			],
-			'three parts, wildcard in third' => [
-				'reference' => 'www.example.*',
-				'expected'  => false,
-			],
-			'three parts, wildcard in first at start with other characters' => [
-				'reference' => '*ww.example.com',
-				'expected'  => false,
-			],
-			'three parts, wildcard in first at end with other characters' => [
-				'reference' => 'ww*.example.com',
-				'expected'  => false,
-			],
-			'three parts, wildcard in first and second' => [
-				'reference' => '*.*.com',
-				'expected'  => false,
-			],
-			'three parts, wildcard in second and last' => [
-				'reference' => 'www.*.*',
-				'expected'  => false,
-			],
-			'three parts, wildcard in first and last' => [
-				'reference' => '*.example.*',
-				'expected'  => false,
-			],
-		];
-	}
-
-	/**
 	 * Tests receiving an exception when an invalid input type is passed as $host.
 	 *
 	 * @dataProvider dataInvalidInputTypeStringable
@@ -540,24 +446,6 @@ final class SslTest extends TestCase {
 	 */
 	public function dataInvalidInputTypeArrayAccess() {
 		return TypeProviderHelper::getAllExcept(TypeProviderHelper::GROUP_ARRAY_ACCESSIBLE);
-	}
-
-	/**
-	 * Tests receiving an exception when an invalid input type is passed.
-	 *
-	 * @dataProvider dataInvalidInputTypeStringable
-	 *
-	 * @covers ::verify_reference_name
-	 *
-	 * @param mixed $input Input data.
-	 *
-	 * @return void
-	 */
-	public function testVerifyReferenceNameInvalidInputType($input) {
-		$this->expectException(InvalidArgument::class);
-		$this->expectExceptionMessage('Argument #1 ($reference) must be of type string|Stringable');
-
-		Ssl::verify_reference_name($input);
 	}
 
 	/**
